chore: Update VSS.css and login.vue
The code changes in VSS.css include commenting out the padding property in the sidebar styling. This change was made to improve the appearance of the sidebar.

In login.vue, the background color of the body element was changed to white and the font family was set to 'Avenir', Helvetica, Arial, sans-serif. These changes were made to enhance the visual design of the login page.